refactor(desktop): move session ui components to feature domain
