+++ apps/desktop/src/App.tsx
@@ -1,11 +1,11 @@
 import { useCallback, useDeferredValue, useEffect, useMemo, useState } from 'react';
 import { AppShell } from '@kay-am/ui';
 import type { SessionId } from '@kay-am/types';
-import { CommandPalette } from './components/CommandPalette';
+import { CommandPalette } from './features/session/components/CommandPalette';
 import { BootSplash } from './components/BootSplash';
 import { ChatView } from './components/chat/ChatView';
 import { ContextPanel } from './features/context/components/ContextPanel';
-import { EndSessionDialog } from './components/EndSessionDialog';
+import { EndSessionDialog } from './features/session/components/EndSessionDialog';
 import { SettingsDialog } from './features/settings/components/SettingsDialog';
 import { ShortcutHelpDialog } from './features/settings/components/ShortcutHelpDialog';
 import { ToastProvider } from './components/Toast';
