+++ apps/desktop/src/__tests__/keyboard/navigation.test.tsx
@@ -74,8 +74,8 @@ import { afterEach, describe, expect, it, vi } from 'vitest';
 import { cleanup, fireEvent, render, screen, within } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import type { Session, SessionId, WorkspaceId } from '@kay-am/types';
-import { EndSessionDialog } from '../../components/EndSessionDialog';
-import { NewSessionDialog } from '../../components/NewSessionDialog';
+import { EndSessionDialog } from '../../features/session/components/EndSessionDialog';
+import { NewSessionDialog } from '../../features/session/components/NewSessionDialog';
 import { SlashCommandPopover } from '../../components/chat/SlashCommandPopover';
 import { ToastProvider } from '../../components/Toast';
