+++ apps/desktop/src/features/session/components/SessionSettingsDialog/index.tsx
@@ -11,12 +11,12 @@ import {
   Trash2,
 } from 'lucide-react';
 import type { SessionId } from '@kay-am/types';
-import { formatError } from '../shared/lib/errors';
-import { useAppStore } from '../store';
-import { SESSION_FEATURES } from '../shared/lib/features';
-import { parseCap } from '../shared/lib/parse-cap';
-import { listLocalBranches, type LocalBranchInfo } from '../features/worktree/worktree';
-import { useToast } from './Toast';
+import { formatError } from '../../../../shared/lib/errors';
+import { useAppStore } from '../../../../store';
+import { SESSION_FEATURES } from '../../../../shared/lib/features';
+import { parseCap } from '../../../../shared/lib/parse-cap';
+import { listLocalBranches, type LocalBranchInfo } from '../../../../features/worktree/worktree';
+import { useToast } from '../../../../components/Toast';
 
 interface SessionSettingsDialogProps {
   sessionId: SessionId;
