+++ apps/desktop/src/features/session/components/StatusBadge/index.tsx
